Proof of concept for user_id based authentication
A user id will be enough to be authenticated with the system. This means
that if an "attacker" gets a hold of a user's id, they can impersonate
him forever.
I have determined that I am ok with this given the type of application
I'm building.
The user_id will be long enough that brute-forcing won't be useful.
QR codes will be used to "automagically" log the user in.