Remove the user_id column from rci_documents
Turns out that having this column here is both redundant and
inconsistent.
It is redundant because the creator of the document is also added to the
lust of acl_owners.
It is incosistent because the user_id field is not actually checked when
we are authorizing. This is because not every protected resource has
such a field.