Making a pivot to a simpler authorization model
I wised up and realized that my makeshift authorization "algorithm" was
too complicated and restrictive.
Additionally, I was making my job harder by trying to make the
application as general-purpose as possible.
Truth is by doing that it was going to be "jack of all trades master of
none" type of app. So I am now embraching the specificity.