Finshed a working skeleton of login page
We have a working skeleton for a login page. Along the way, I learned
about data-binding and how that might be implemented in javascript.
Seems like it could be useful, but I'll hold off. I want to build the
skeleton for a few other pages so I can figure out the common things I
keep doing.