Implement the observable pattern
Played around with the observable pattern and found it to be useful.
For now I applied it to the http requests.

The goal is not to shy away from the asynchronous nature of javascript
by trying to mold it and make it look synchrnous (i'm looking at you
"promises"). There is still some work to be done in figuring out how to
arrange the code in a way that looks more declarative.