Implement res_life_staff rci list
And with this i'll take a break on sketching the functionality to
actually filling it in.
I'm choosing to stop here because I have written a decent amount of code
so far. Cleanup and refactoring are due, especially since I have a
better idea of how the rest of the project is going to be structured

Sketching any further will get harder since the foundations are still
just sketches