Modify the full-text-search tokenizer
Testing has revealed that weird things happen if the search string is
not quoted.
Additionally, it looks like the hyphen and underscore characters were
not being recognized as tokens, which should be the case ...i think?
For now things work better, but i might need to revisit search strategy
later.