Streamling dashboard authorization flow
The main thing here is to move to exclusively using permissions to
authorize instead of role names.
The role names are arbitrary and schools should be free to create as
many as they want with whatever name.
The job of the system is to react accordingly and display the correct
thing for a user solely based on their permissions