Setup plumbing for system settings
Most of our templates can be viewed by users with varying permissions.
What is actually displayed might differ, but it is usually the same
template.
The system settings template is different. Only an administrator should
be even able to get here. So I implemented a admin_required decorator to
have that happen.